added test_chat.py to test the chat function (just creates UI class and calls it)
