+++ test/test_chat.py
@@ -1,0 +1,19 @@
+from PyQt5.QtGui import *
+from PyQt5.QtCore import *
+from PyQt5.QtWidgets import *
+
+import chat
+
+# @desc
+# initializes all UI widgets
+class UI:
+    def __init__(self):
+        self.qapp = QApplication(sys.argv)
+        self.window = QMainWindow()
+        self.main_widget = chat.MainWidget()
+        self.window.setCentralWidget(self.main_widget)
+        self.window.show()
+        sys.exit(self.qapp.exec_())
+
+if __name__ == '__main__':
+    someUI = UI()
